fix(analytics-control-bar): moved analytics control bar utilities into its own component at request of gus
