+++ src/components/analytics-control-bar-space-filter/index.tsx
@@ -16,7 +16,7 @@ import {
   BackButton,
   FilterDeleteButton,
   SubmitButton,
-} from '../analytics-control-bar';
+} from '../analytics-control-bar-utilities';
 
 import {
   AppBar,
